[DBInstance] Rework error handling tests (#279)
This commit performs a deep rework of the error handling test code. In
particular, the commit adds a new method called
`test_handleRequest_error` and introduces a set of parametrized tests
exercising distinct RDS API exception classes and error codes. The error
codes and classes are provided in a declarative way (rather than an
independent test method per single error code/exception.

The motivation for this change is to encourage a more exhaustive test
coverage on distinct error codes/classes. With the current
implementation adding a new test for an error code/class requires
writing a new test method with a minimal bootstrap and validation. The
new approach allows one to add a new entry into a static mapping of
ErrorCode -> HandlerErrorCode and the bootstrap would be performed
automatically.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>