[DBParameterGroup] Filtered DescribeDBParameters
This commit adds an additional filtering to DescribeDBParameters.
Having it in-place reduces the page depth of the cluster parameter
fetch. The amount of parameters to be fetched is limited to the
parameters that are directly enumerated in the ResourceModel.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Relevant PR: https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-rds/pull/284

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>