[DBInstance][DBCluster] Handle StorageTypeNotSupportedException
This commit adds `StorageTypeNotSupportedException` to the
InvalidRequest mapping. A recent attempt to address this exception was
insufficient
(https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-rds/pull/278)
as both `StorageTypeNotSupportedException` and
`StorageTypeNotSupportedFault` could be returned by the RDS API
depending on the implementation.

This commit adds both error class and the error code to the default
error rulesets of DBInstance and DBCluster.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>