[Common] Split cfn-common and cfn-test-common (#314)
This commit splits up aws-rds-cfn-common package in 2. The test helper
suite is moved to a new package called aws-rds-cfn-test-common. The
motivation for this change is to break down the common dependencies into
compile and test scopes. The old approach forced us to include test
dependencies into the compile scope, it is eliminated with this commit.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>