[TestCommon] Implement access permission verifier
This commit adds a mechanism to programatically verify the schema
permission integrity. The motivation for this change is to close the gap
between the schema definition and the actual implementation. As practice
indicates, one of the most common problems occurs when a new API call is
being added and no corresponding access permission is requested by the
schema.

The approach in a nutshell is to inspect the corresponding client object
after every unit test execution and verify that all invocations were
declared in the corresponding schema access permissions.

The public API for access permission verification is:

    verifyFASPermissions(clientMock);

It is supposed to be called upon every unit test completion, e.g:

    @AfterEach
    public void tearDown() {
        verifyFASPermissions(clientMock);
    }

The verifier will throw a MockitoAssertionError if the permission is
missing.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>