[DBInstance] Move status definitions to enums (#322)
* [DBInstance] Move status definitions to enums

This commits performs a shallow refactoring of the status tracking layer
which was implemented as a trivial set of string constants.
DBInstanceStatus was the first case where we had to attach some
additional meta-information to the status (terminal status flag).

Given our ultimate aim to let the stack expose unrecoverable problems as
soon as possible, we are adding the same bit of information (terminal
status flag) to OptionGroupStatus. This move motivates an introduction
of a new inheritance chain.

Firstly, we introduce a `Status` interface which defines a minimalistic
API for status-like enums. Whereas enums is a great way of handling
discrete values, java adds some limitations on using the actual
inheritance on these classes, hence we favor interfaces.

The second family grid or status-like objects is TerminableStatus. This
interface declares a basic API to check whether this state is transient
or terminal.

Other constant-defined statuses were moved to the same package and
gained the same implementation of the base methods. Using this enum
package is encouraged instead of const-based status handling.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>