[DBCluster] Stabilize on pending actions
This commit adds an additional condition to the DBCluster stabilization
logic: apart from checking that the cluster has achieved an Available
state, it also checks there are no pending actions scheduled for an
immediate execution.

The motivation for this change is the RDS API specifics: whereas some
attributes are being applied immediately, the rest could be scheduled up
for a delayed execution and picked up as soon as possible. This is the
case for the following attributes:
  - masterUserPassword
  - iamDatabaseAuthenticationEnabled
  - engineVersion

This commit ensures the handler will inspect pendingActions before
moving on.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>