[DBInstance] Set engineVersion conditionally (#326)
This commit changes the way Translator constructs a ModifyDBInstance
request. In particular, it only sets `engineVersion` attribute if there
was any change. The motivation for this change is to avoid a potential
immediate engine version upgrade that could be triggered by the API
when the request contains a non-null `engineVersion` and marked for an
immediate execution (always the case for CFN requests).

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>