[Common] Rework runtime config system
This commit introduces a new config system called `RuntimeConfig` as a
replacement of `HandlerConfig`.

The motivation for the change is to move away from the hardcoded config
to a local file-backed extendable system.

As the primary config data medium we chose plain old properties files.
It is not the best way to store structured data, but works well enough
for flat primitive-only configs.

A resource using RuntimeConfig is expected to expose a new config
properties folder and add it to the compiled classpath.

The default config sets are:
- config/runtime.properties
- config/test.properties