[DBClusterParameterGroup] Do not describe parameters if empty
This commit is a follow-up on https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-rds/pull/337.

The motivation for the change is to eliminate a describe parameters call
completely if the model parameter set is empty. Under the hood, this
change is dictated by the fact that an empty parameter set is an invalid
request from RDS API's perspective. In this case we should choose
between these 2 options:
* set no filters of the param set is empty
* do not describe current parameters at all (as the upcoming validation
  would cross-match an empty parameter set against a presumably empty
  default parameter set, hence it is a no-op).

This commit is a move from point 1 to point 2.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>