[DBCluster] Unset default port on Update
This commit ensures the UpdateHandler unsets the default port on update.
The motivation for this change is to eliminate cases where RDS API would
reject a modify request due to a non-null port value. This behavior is
engine-specific hence setting port to null upon a no-change seems to be
the one-size-fits all aproach.

This commit is a follow-up on a change we delivered earlier:
https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-rds/pull/343

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>