[DBInstance] Check events on stabilization
This commit introduces an additional check that would be performed upon
a DBInstance stabilization. The motivation for this change is to extend
a limited set of criterias in use to declare an operation completion and
its status.

In particular, using events could be the only way to detect a mutation
issue when RDS API indicates that the initial request returns a
successful response and the resource eventually reaches an `available`
status, despite the fact that the mutation failed and only left a
corresponding trace in the event stream.

Using events in stabilization requires an additional permission:
rds:DescribeEvents. Given this permission has never been a part of the
DBInstance permission set, we wrap it in a silent try-catch block so to
avoid turning events into a hard dependency (at least, for now).

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>