[DBInstance] Modify with a lookback at the current state after create
This commit changes the way CreateHandler performs an update after
create. Historically, not every attribute of DBInstance could be set
with a
CreateDBInstance/RestoreDBInstanceFromDBSnapshot/CreateDBInstanceReadReplica
request. For this reason the handler would emit a follow-up
ModifyDBInstance request that would ensure the right attribute set.
Since DifferenceUtils lib was introduced, Translator started assembling
minimalistic modify requests, only including the attributes that are
about to be changed. Given that a CreateHandler request contains no
previous resource state, we were using a barebone model builder and
feed it to diff. This gives an all-changed modify request. This commit
alters this approach and describes a DBInstance right before modifying
it. Using this technique allows us to emit minimalistic modify requests.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>