[DBInstance] Move modify-after-create to a standalone method (#359)
This commit moves modify-after-create Translator to a separate method.
The motivation for this change is to get rid of an overly-complicated
logic in the current implementation of
`Translator.modifyDbInstanceRequest`. There is a significant difference
between a regular update and an update after create that is only
supposed to concern a limited set of attributes. An example of this
logic is handling iops or allocatedStorage parameters.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>