[All] Bump AWS SDK version to 2.18.24
This commit bumps all package aws sdk versions to `2.18.24`. The
motivation for the change is a potential regression introduced by a
recent SDK release.

The test regression was noticed in
https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-rds/pull/353/files#diff-608a303a115402d9a7b63d5e9a01808ea8b66ea1508f49e6d5d4115354c4fa83R94-R95
and caused a failure in `RdsClientProviderTest` probing the presence of
the provided version in the RDS API request body.

The SDK was patched and re-released, hence we are reverting the disabled
test.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>