[DBInstance] Unset AllocatedStorage and Iops on rollback
This commit changes the handler behavior on a rollback update.
The current strategy is to ensure AllocatedStorage and Iops
changeability. Effectively, this means that the handler would ensure
that neither of these attributes is being reduced upon a rollback. An
attempt to reduce the corresponding attribute is known to be rejected by
the RDS API, hence the handler would keep the old (pre-rollback) value.
This commit alters this strategy to providing no pre-rollback values at
all if either of these attributes is identifier as non-updatable. The
translator would provide a null-value instead.

The motivation behind this change is to expand the applicability of the
"keep whatever is working" on a case when the initial update failed.
The existing approach would only work if the initial change was applied
successfully and failure happened somewhere down the stack tree. The
change would be invalid if the initial update never went through.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>