[Common] Implement Drift Detector Test Helper
This commit implements a new test helper called ResourceDriftTestHelper.
This is an assertion helper that checks if there is a drift between an
input and an output, based on the resource schema definition provided.

In particular, the drift detector would engage propertyTransform
instructions from the schema and compare the transform formula result
with the output. If neither the original input value nor any of the
transformation outputs match the output, the resource is considered to
have drifted.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>