[Common] Handle SdkServiceException
This commit adds another SDK-related exception class to the default
error rule set. The semantics of SdkServiceException is pretty similar
to SdkClientException, with the difference that the issue occurred on
the server side (e.g. backend unavailable).

The exception is being handled as an internal service error and hence
would be retried.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>