[DBInstance] Mark DeleteAutomatedBackups as WriteOnly
This commit marks `DeleteAutomatedBackups` as a write-only property. The
morivation for this change is to align this attribute behavior with its
purpose: the flag is semi-virtual and is only used upon a delete. Up
until this point it is an ephemeural attribute that is not stored
anywhere, hence bugging CFN users with false drift detector alarms.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>