[DBInstance] Mark AllowMajorVersionUpgrade as WriteOnly
This commit marks `AllowMajorVersionUpgrade` as a write-only property.
The motivation is similar to
https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-rds/pull/401
in terms of the nature of the attribute: the RDS API won't accept it
upon a create and won't serve it back with a describe-db-instances. The
flag is only used upon a modify request to indicate that a major engine
version could be changed.

Refs https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-rds/issues/400
Fixes https://github.com/aws-cloudformation/cloudformation-coverage-roadmap/issues/1511

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>