[DBClusterParameterGroup] ReadHandler Includes Parameters (#414)
This commit adds `parameters` attribute to the response of the
`ReadHandler`. The current implementation marks this attribute as
write-only, which disqualifies the attribute for a drift detection.

As a side effect, this code change includes a deep refactoring in this
handler unit test suite. The refactoring promotes a wide use of the
standard testing utility `test_handleRequest_base`.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>