[OptionGroup] Fix drift on OptionConfigurations
This commit addresses a false drift detection on OptionConfiguration
items. The nature of this drift comes from the RDS API which populates
OptionConfiguration objects with default OptionSettings. I.e. if the
stack defines only a few OptionSettings within an OptionConfiguration
object, RDS would populate it with all default values. CFN looses track
of these settings and indicates the drift.

The idea behind this change is to hide all settings that are set to
defaults. This could cause a counter-drift when the stack defines a set
of OptionSettings that happened to match the default values. In this
case it is an equivalent of not defining these properties at all, hence
we consider this inversion to be a reasonable way to address the initial
drift.

Refs https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-rds/issues/400.
Fixes https://github.com/aws-cloudformation/cloudformation-coverage-roadmap/issues/1310.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>