[DBInstance] Set DeletionProtection after create
This commit instructs Create handler to modify a creating instance if
DeletionProtection is set to true. The reason for this change is to
compensate for a non-unified way distinct APIs create instance APIs
handle this flag. An explicit set after create would ensure no false
drift is detected on this attrubute.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>