[DBInstance] Set performance insights key on modify-after-create
This commit changes ModifyAfterCreate method of the Translator class.
The method is instructed to pass the performance insights(PI) kms key ID if
PI is enabled. RDS API expects the key ID upon every modify request if the
instance was created with PI enabled and a key ID was provided.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>