[DBInstance] Set source instance engine on read replica
This commit adds an additional source db instance fetch upon a read
replica create. At the moment, storage-related attributes are being
set upon a create for non-SQLServer engines only. Due to the safeguard
in place, the handler treats all unset engine types as SQLServer and
hence sets no storage-related attributes on read replicas.

The idea is to set the replica engine type from the source instance so
the handler doesn't have to make a guess whether it is a SQLServer
engine.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>