[DBInstance] Handle a restore from snapshot when MultiAZ=false
This commit fixes a faulty behavior on a corner case when an instance is
being restored from a snapshot and its engine is not explicitly set but
MultiAZ is set to false. The current implementation instructs the
handler to fetch the snapshot only if MultiAZ is unset. The engine data
is used later on to detect the required storage type modifications.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>