[DBCluster][DBInstance] Fix empty feature name in AssiciatedRoles
This commit addresses a role mutate stabilization issue that occurs if
the role association feature name is an empty string. RDS API treats this as a
no-value and returns an null feature name back. Both DBInstance and
DBCluster handlers stabilize on both RoleARN and FeatureName attributes
upon an addition, hence this discrepancy between an empty string and a
null value cause an infinite stabilization.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>