[DBInstance] Mark UseDefaultProcessorFeatures as write-only
This commit changes the visibility of `UseDefaultProcessorFeatures`
attribute to WriteOnly. The motivation for this change is to get aligned
with the SDK which does not expose the attribute for the describe API
interface. I.e. the attribute is write-only from the API perspective,
hence keeping it as read-write would cause a false drift.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>