[Common] Implement drift detector
This commit implements a local drift detector logic.
The implementation was used in the testing code previously. The change
introduces a new class called DriftDetector. The class implements a
logic of a recursive descending pairwise comparison. The equality check
takes into account the property transformation logic declared in the
resource schema.

Signed-off-by: Oleg Sidorov <sidorovo@amazon.com>