Add EngineLifecycleSupport integration to Global Clusters
Global clusters can be created with the EngineLifecycleSupport
field via CloudFormation. All subsequent updates to this field
on the resource will be blocked to prevent unintended customer
resource replacement. EngineLifecycleSupport is marked as
"Some Interruption" causing to preserve future compatibility for
when a formal modify option is available for this field.