fix(DBInstance): allow deleting cluster instances through CCAPI
This change fixes an issue preventing cluster instances from being
deleted using Cloud Control API.

Final snapshots are not possible for read replicas and cluster
instances. The issue was that the handler was determining this based on
instance details provided in the request. CCAPI does not provide these
details. The fix is to retrieve the instance details from RDS instead of
relying on the request.

Test: unit tests and contract tests.

Issue: aws-cloudformation/cloudformation-coverage-roadmap#2013